Add back featured + tests.
